Stay inside the current URL scheme
If the web server is configured to serve pages over both http and
https, then previously the full URL configured as cfg_gitweb_url
and cfg_browsercgi_url would have been used.  Unfortunately this
is guaranteed to break the scheme that is not embedded into the
url when trying to use git-browser.

Instead replace the scheme, host and port portion of those URLs
with the value from the current location thus allowing git-browser
to work for both schemes simultaneously.