git-browser.cgi: fix handling of latin-1 characters
Handle commit characters the same way gitweb.perl does by
attempting to use UTF-8 encoding for them and if that's not
valid then using the ISO-8859-1 encoding for them.

This is done on a line-by-line basis so that, for example,
the author name may end up being treated as ISO-8859-1
while the commit message is treated as UTF-8.