by_commit.html: avoid vertical drawing drift
The font size is set to 11px explicitly and the cell border for
each line is 1px thereby accounting for 13px of the distance between
each row.

The height, however, was previously set to 1.2em and the line-height
left unspecified.

The problem is that 11px * 1.2 = 13.2.  This means that including
the 1px border on top and bottom, each row of text is 15.2px high.

On some browsers that means that while each line of text will
actually advance 15.2px (meaning that after advancing 3 lines you've
moved 45.6px which rounds to 46 NOT 45), the line drawing will only
advance exactly 15px per row.

This means that starting with the third row of text and every fifth
line thereafter, the text of the rows will be positioned vertically
one more pixel farther down the page than it should be.

Fix this by explicitly making the height and line-height exactly
13px so that there's no fractional accumulation going on to gradually
push the text more and more out-of-sync as you move down the page.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>