git-browser.cgi: include all tags in default graph
The default graph shows all heads, but it does not necessarily show
all tags unless all the tags are reachable from all the heads.

Fix that by explicitly including all tags in the default graph so
that all tags can be viewed even if they are not all reachable from
the full set of heads.

Signed-off-by: Kyle J. McKay <mackyle@gmail.com>